move support for suggested fixes to sema
